rename: graph->envs
