+++ tests/envs/test_hex_grid.py
@@ -1,6 +1,6 @@
 import unittest
 import numpy as np
-from w9_pathfinding.graph import HexGrid, HexLayout
+from w9_pathfinding.envs import HexGrid, HexLayout
 
 
 class TestHexGrid(unittest.TestCase):
