+++ tests/pf/test_graph_pf.py
@@ -1,5 +1,5 @@
 import unittest
-from w9_pathfinding.graph import Graph
+from w9_pathfinding.envs import Graph
 from w9_pathfinding.pf import DFS, BFS, BiBFS, Dijkstra, BiDijkstra #, SpaceTimeAStar
 
 SHORTEST_PATH_ALGORITHMS = [Dijkstra, BiDijkstra] #, SpaceTimeAStar]
