+++ w9_pathfinding/pf.pyx
@@ -1,7 +1,7 @@
 # distutils: language = c++
 
 from w9_pathfinding cimport cdefs
-from w9_pathfinding.graph cimport _AbsGrid, _AbsGraph, Graph, ReservationTable
+from w9_pathfinding.envs cimport _AbsGrid, _AbsGraph, Graph, ReservationTable
 
 
 def to_node_id(graph, node):
